Move code to `features/project` directory
