+++ web-common/src/features/project/selectors.ts
@@ -1,0 +1,13 @@
+import { parseDocument } from "yaml";
+import { createRuntimeServiceGetFile } from "../../runtime-client";
+
+export function useProjectTitle(instanceId: string) {
+  return createRuntimeServiceGetFile(instanceId, "rill.yaml", {
+    query: {
+      select: (data) => {
+        const projectData = parseDocument(data.blob)?.toJS();
+        return projectData?.title ?? projectData?.name;
+      },
+    },
+  });
+}
