+++ web-common/src/features/project/shorthand-title/index.ts
